feat: Rearrange menu elements and nix the lineyness.
